KVM: PPC: Book3S HV: Make physical thread 0 do the MMU switching
We have been seeing evidence of occasional memory corruption in HV
guests, as well as sporadic machine check interrupts with SRR1 indicating
duplicate TLB entries.  This was bisected down to the patch "KVM: PPC:
Book3S HV: Align physical CPU thread numbers with virtual", which was
part of the POWER8 enablement patch set.  One of the things that patch
did was to move the responsibility for setting the MMU-related SPRs from
physical thread 0 to the thread of whichever vcpu initiated the entry
to the guest.

The main effect of this patch is to move that responsibility back to
thread 0.  That means that if the vcpu on some other thread wants to
enter the guest, thread 0 has to go through the guest entry code as far
as the MMU switch and then go into nap mode.  If thread 0's vcpu wants
to start running later, we need to wake up thread 0 from nap and have
it load up the guest registers and jump into the guest.  Alternatively,
if some thread needs to exit the guest, we need to wake up thread 0 and
have it do the MMU switch back to the host context.

Doing this eliminates the machine check interrupts and memory
corruption.

The diff is large because some blocks of code got moved.  Specifically,
the code that loads the guest MMU context got moved to before we load up
the guest registers (including transactional state), and the code that
loads the host MMU context got moved to after we save the guest register
state.

Since the master or primary thread is once again always thread 0, we
no longer need the 'is_master' field of struct kvm_vcpu_arch.

Signed-off-by: Paul Mackerras <paulus@samba.org>
(cherry picked from commit 8ebba8854bb64973c8a3ca30da49836976deb6a0)