KVM: PPC: Book3S HV: Send some subcommands of H_SET_MODE to userspace
This removes the code that handles the H_SET_MODE_RESOURCE_LE and
H_SET_MODE_RESOURCE_ADDR_TRANS_MODE subfunctions of the H_SET_MODE
hypercall from the kernel.  Instead we now return H_TOO_HARD which
causes the hypercall to be sent up to userspace to be handled there.
In addition we now also send any other subfunction which we don't
recognize to userspace.

The reason for doing these two subfunctions in userspace is that they
need to modify LPCR across all vcpus of the guest.  Modifying LPCR in
the kernel like this introduces a race between the kernel's
modification and any modification that userspace might be doing on
another vcpu.  Therefore it's better to let userspace do all the
modifications, so it can do any necessary synchronization itself.

This also adds code to make sure that the MSR_LE bit in intr_msr
(the MSR value we set when synthesizing an interrupt for the guest)
is in sync with the ILE bit in the virtual core's LPCR value.  This
is necessary for implementing the LE subfunction of H_SET_MODE in
userspace.

Signed-off-by: Paul Mackerras <paulus@samba.org>