powerpc/pseries: Disable relocation on exception while going down during crash.
Disable relocation on exception while going down even in kdump case. This
is because we are about clear htab mappings while kexec-ing into kdump
kernel and we may run into issues if we still have AIL ON.

This is at least part of the fix for kdump failures that we are seeing
in bug 103693.

Signed-off-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
(cherry picked from commit 3ec8b78fcc5aa7745026d8d85a4e9ab52c922765)
Signed-off-by: Paul Mackerras <paulus@samba.org>