cpuidle/powernv: Enable Fastsleep at boot time
Fast sleep can be enabled today, only after writing into the proc interface
/proc/sys/kernel/powersave-nap with a value greater than 1. Remove this
constraint, now that we have a stable framework to support fast sleep, so
that it is enabled by default at boot.

However the same proc interface is also used to convey if deep idle states
beyond snooze can be entered into or not. Hence retain the check on
powersave-nap in fast sleep to verify if this is the case.

Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>