KVM: PPC: Book3S HV: Move guest TM restore code out of PMU restore sequence
Somehow, the code that restores the guest transactional memory state
got put in the middle of the code sequence that restores the guest
PMU (performance monitor unit) state.  This results in corruption of
the value written to MMCR0 if the guest is in transactional state.

This fixes it by moving the TM state-restoring code to come just before
the PMU state-restoring code.  This comes out in the patch as the
first part of the PMU state-restoring code being moved down to just
before the second part of the PMU state-restoring code.

Signed-off-by: Paul Mackerras <paulus@samba.org>