KVM: PPC: Book3S HV: Save/restore host PMU registers that are new in POWER8
Currently we save the host PMU configuration, counter values, etc.,
when entering a guest, and restore it on return from the guest.
(We have to do this because the guest has control of the PMU while
it is executing.)  However, we missed saving/restoring the SIAR and
SDAR registers, as well as the registers which are new on POWER8,
namely SIER and MMCR2.

This adds code to save the values of these registers when entering
the guest and restore them on exit.  This also works around the bug
where setting PMAE with a counter already negative doesn't generate
an interrupt.  This was already worked around for the guest PMU state
in an earlier commit, and is worked around for the host PMU state here.

Signed-off-by: Paul Mackerras <paulus@samba.org>