powernv, cpufreq: Export nominal frequency via sysfs.
Create a driver attribute named "cpuinfo_nominal_freq" which will in
turn create a read-only sysfs interface that will be used to export
the nominal frequency to the userspace. This will be necessary for
creating an optimal "performance" policy which should be running the
on-demand governor with "scaling_max_freq" to be set to the value
exported via "cpuinfo_max_freq" and "scaling_min_freq" to be set to
the nominal frequency exported via "cpuinfo_nominal_freq".

The patch caches the values of max, min, nominal pstate ids and
nr_pstates queried from the DT during the initialization of the driver
so that they can be used in other places in the driver for
validatation.

Also, it adds a helper method that returns the frequency corresponding to
a pstate id.

This has been backported from the version posted against mainline
which can be found here:
https://www.mail-archive.com/linuxppc-dev@lists.ozlabs.org/msg76990.html

Reviewed-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>