kvm: Set the runlatch bit of a CPU just before starting guest
The secondary threads in the core have their runlatch bits cleared since they
are offline. When the secondary threads are called in to start a guest their
runlatch bits need to be set to indicate that they are busy. The primary
thread has its runlatch bit set though, but there is no harm in setting this
bit once again. Hence set the runlatch bit for all threads before they start
guest.

Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Acked-by: Paul Mackerras <paulus@samba.org>
Reviewed-by: Srivatsa S. Bhat <srivatsa.bhat@linux.vnet.ibm.com>