powerpc: Don't try to set LPCR unless we're in hypervisor mode
Commit c7062d83fe7b ("powerpc/ppc64: Do not turn AIL (reloc-on
interrupts) too early") added code to set the AIL bit in the LPCR
without checking whether the kernel is running in hypervisor mode.
The result is that when the kernel is running as a guest (i.e.,
under PowerKVM or PowerVM), the processor takes a privileged
instruction interrupt at that point, causing a panic.  The visible
result is that the kernel hangs after printing "returning from
prom_init".

This fixes it by checking for hypervisor mode being available
before setting LPCR.  If we are not in hypervisor mode, we enable
relocation-on interrupts later in pSeries_setup_arch using the
H_SET_MODE hcall.

This fixes BZ 108728.

Signed-off-by: Paul Mackerras <paulus@samba.org>