KVM: PPC: Book3S PR: Unimplemented SPRs in supervisor mode don't cause trap
The Power ISA states that an mtspr or mfspr to/from an unimplemented
SPR should be a no-op in privileged mode, rather than causing an
program interrupt (0x700 vector), with the exception of mtspr to SPR 0
and mfspr from SPRs 0, 4, 5 or 6.

Currently our SPR emulation code doesn't follow this rule.  This
modifies the code in kvmppc_core_emulate_m[ft]spr_pr() to check
the PR bit in the MSR when we detect an unknown SPR number, and
only return EMULATE_FAIL (which results in a program interrupt)
if PR is 0 or the SPR number is one of the ones which are specifically
defined to cause a program interrupt.

Signed-off-by: Paul Mackerras <paulus@samba.org>