KVM: PPC: Book3S PR: Handle facility-unavailable interrupts gracefully
At present, if a PR guest on a POWER8 machine tries to access some
disabled functionality such as transactional memory, the result is
a facility-unavailable interrupt, which isn't handled in
kvmppc_handle_exit_pr(), resulting in a call to BUG(), crashing
the PR host kernel.

This adds code to handle the facility-unavailable interrupts and
give the guest an illegal instruction interrupt, instead of crashing
the PR host.

Signed-off-by: Paul Mackerras <paulus@samba.org>