tick-broadcast/cpuidle: Fix the demuxing of PPC_MSG_TIMER IPI message
The PPC_MSG_TIMER IPI message slot was introduced for the tick broadcast
IPIs which are required to wakeup sleeping CPUs. The decrementer of the
CPUs that enter fast sleep stops as a consequence of entering the idle
state. Therefore such CPUs have to be woken up in time to handle their
timers by a broadcast CPU which sends the PPC_MSG_TIMER IPIs to them.

This IPI message is being parsed wrongly in smp_ipi_demux(). Thus the
tick broadcast interrupt handler is never executed on the sleeping CPU.
This could have led to unpleasant side effects like not handling timers
in time on the sleeping cpus. But since the sleeping CPUs still receive
the tick broadcast IPI, they are awoken from the idle state and their
decrementers are back in action.
  As a result, its possible that they are managing to handle timers
before they go to sleep again.

Hence timers are being handled on the sleeping cpus although the tick
broadcast interrupt handler, which is actually supposed to ensure that
is never being called today due to the wrong number of shift bits while
parsing the tick broadcast IPI.
  However we need to note that as a result of this discrepency, timer
handling on the sleeping cpus may be unstable. This could be one of
the reasons we are observing some softlockups in the cpuidle wakeup path.

Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>