tick-broadcast/cpuidle: Fix the programming of the broadcast hrtimer
Today CPUs in fast sleep are being woken up to handle their timers
by the tick broadcast framework using a hrtimer queued on a nominated
broadcast CPU. The hrtimer is programmed for the earlier of the next
wakeup and a broadcast period which happens to be a jiffy. This
programming is being done incorrectly today. The current time
is noted, the tick broadcast interrupt handler is called, then the
time at which the hrtimer needs to be programmed is decided. By
then the noted current time would be stale and the hrtimer will
be forward much ahead than required, leading to delayed broadcast
interrupts being delivered to sleeping cpus.

Fix this by noting the current time just before programming the hrtimer.

Signed-off-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>