powerpc/eeh: Report frozen parent PE prior to child PE
When we have the corner case of frozen parent and child PE at the
same time, we have to handle the frozen parent PE prior to the
child. Without clearning the frozen state on parent PE, the child
PE can't be recovered successfully.

There're 2 ways (polling and interrupt) to have frozen PE to be
reported. If we have frozen parent PE out there, we have to report
and handle that firstly.

It's responsing to bug#109562.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>